devServer updates & move component-specific props into nested config object
