+++ npm/react/examples/a11y/cypress.config.js
@@ -4,21 +4,15 @@ const { defineConfig } = require('cypress')
 
 // load file devServer that comes with this plugin
 // https://github.com/bahmutov/cypress-react-unit-test#install
-const devServer = require('@cypress/react/plugins/react-scripts')
+const { devServer } = require('@cypress/react/plugins/react-scripts')
 
 module.exports = defineConfig({
   video: false,
   fixturesFolder: false,
-  testFiles: '**/*spec.js',
   viewportWidth: 500,
   viewportHeight: 500,
   component: {
-    setupNodeEvents (on, config) {
-      devServer(on, config)
-
-      // IMPORTANT to return the config object
-      // with the any changed environment variables
-      return config
-    },
+    devServer,
+    testFiles: '**/*spec.js',
   },
 })
