+++ npm/react/examples/find-webpack/cypress.config.js
@@ -2,28 +2,21 @@ const { defineConfig } = require('cypress')
 
 // @ts-check
 const findReactScriptsWebpackConfig = require('@cypress/react/plugins/react-scripts/findReactScriptsWebpackConfig')
-const { startDevServer } = require('@cypress/webpack-dev-server')
-const _ = require('lodash')
+const { devServer } = require('@cypress/webpack-dev-server')
 
 module.exports = defineConfig({
   video: true,
   projectId: 'jq5xpp',
-  component: {
-    testFiles: '**/*.spec.{js,ts,jsx,tsx}',
-    componentFolder: 'src',
-  },
   env: {
     'cypress-react-selector': {
       root: '#__cy_root',
     },
   },
   component: {
-    setupNodeEvents (on, config) {
-      const map = _.map([4, 8], (n) => n * 2)
-
-      console.log(map)
-      require('@cypress/code-coverage/task')(on, config)
-      const webpackConfig = findReactScriptsWebpackConfig(config)
+    componentFolder: 'src',
+    testFiles: '**/*.spec.{js,ts,jsx,tsx}',
+    devServer(cypressDevServerConfig) {
+      const webpackConfig = findReactScriptsWebpackConfig(cypressDevServerConfig)
 
       const rules = webpackConfig.module.rules.find((rule) => !!rule.oneOf).oneOf
       const babelRule = rules.find(
@@ -32,17 +25,16 @@ module.exports = defineConfig({
         },
       )
 
-      typeof babelRule.options !== 'string' &&
+      if (typeof babelRule.options !== 'string') {
         babelRule.options.plugins.push(
           require.resolve('babel-plugin-istanbul'),
         )
+      }
 
-      on('dev-server:start', (options) => {
-        return startDevServer({
-          options,
-          webpackConfig,
-        })
-      })
+      return devServer(cypressDevServerConfig, { webpackConfig })
+    },
+    setupNodeEvents (on, config) {
+      require('@cypress/code-coverage/task')(on, config)
 
       // IMPORTANT to return the config object
       // with the any changed environment variables
