+++ npm/react/examples/nextjs-webpack-5/cypress.config.js
@@ -2,9 +2,11 @@ const { defineConfig } = require('cypress')
 
 module.exports = defineConfig({
   video: false,
-  testFiles: '**/*.spec.{js,jsx}',
   viewportWidth: 500,
   viewportHeight: 800,
-  componentFolder: 'cypress/components',
   pluginsFile: 'cypress/plugins.js',
+  component: {
+    componentFolder: 'cypress/components',
+    testFiles: '**/*.spec.{js,jsx}',
+  }
 })
