+++ npm/react/examples/react-scripts-typescript/cypress.config.ts
@@ -1,18 +1,13 @@
 import { defineConfig } from 'cypress'
-
-const devServer = require('@cypress/react/plugins/react-scripts')
+const { devServer } = require('@cypress/react/plugins/react-scripts')
 
 export default defineConfig({
   video: false,
-  testFiles: '**/*cy-spec.tsx',
   viewportWidth: 500,
   viewportHeight: 800,
-  componentFolder: 'src',
-  e2e: {
-    setupNodeEvents (on, config) {
-      devServer(on, config)
-
-      return config
-    },
+  component: {
+    devServer,
+    componentFolder: 'src',
+    testFiles: '**/*cy-spec.tsx',
   },
 })
