+++ npm/react/examples/react-scripts/cypress.config.js
@@ -4,22 +4,16 @@ const { defineConfig } = require('cypress')
 
 // load file devServer that comes with this plugin
 // https://github.com/bahmutov/cypress-react-unit-test#install
-const devServer = require('@cypress/react/plugins/react-scripts')
+const { devServer } = require('@cypress/react/plugins/react-scripts')
 
 module.exports = defineConfig({
   video: false,
-  testFiles: '**/*cy-spec.js',
   viewportWidth: 500,
   viewportHeight: 800,
   experimentalFetchPolyfill: true,
-  componentFolder: 'src',
   component: {
-    setupNodeEvents (on, config) {
-      devServer(on, config)
-
-      // IMPORTANT to return the config object
-      // with the any changed environment variables
-      return config
-    },
+    devServer,
+    componentFolder: 'src',
+    testFiles: '**/*cy-spec.js',
   },
 })
