+++ npm/react/examples/visual-testing-with-applitools/cypress.config.js
@@ -6,26 +6,24 @@ const { defineConfig } = require('cypress')
 // https://github.com/bahmutov/cypress-react-unit-test#install
 const devServer = require('@cypress/react/plugins/react-scripts')
 
- module.exports = defineConfig({
-  "video": false,
-  "fixturesFolder": false,
-  "testFiles": "**/*spec.js",
-  "viewportWidth": 1000,
-  "viewportHeight": 1000,
-  "componentFolder": "src",
-  "env": {
-    "coverage": false
+const getApplitoolsWrappedSetupFn = (exports) => {
+  const placeholderModule = { exports }
+  require('@applitools/eyes-cypress')(placeholderModule)
+  return placeholderModule.exports
+}
+
+module.exports = defineConfig({
+  video: false,
+  fixturesFolder: false,
+  viewportWidth: 1000,
+  viewportHeight: 1000,
+  env: {
+    coverage: false
   },
   component: {
-		setupNodeEvents(on, config) {
-      devServer(on, config)
-
-       // IMPORTANT to return the config object
-       // with the any changed environment variables
-       return config
-		}
-  }
- })
-
-// @ts-ignore
-require('@applitools/eyes-cypress')(module)
+    devServer,
+    setupNodeEvents: getApplitoolsWrappedSetupFn((on, config) => {
+      // component node events setup code
+    })
+  },
+})
