+++ npm/react/examples/visual-testing-with-happo/cypress.config.js
@@ -4,21 +4,21 @@ const { defineConfig } = require('cypress')
 
 // load file devServer that comes with this plugin
 // https://github.com/bahmutov/cypress-react-unit-test#install
-const devServer = require('@cypress/react/plugins/react-scripts')
+const { devServer } = require('@cypress/react/plugins/react-scripts')
 
 // @ts-ignore
 const happoTask = require('happo-cypress/task')
 
 module.exports = defineConfig({
   video: false,
-  testFiles: '**/*cy-spec.js',
   viewportWidth: 400,
   viewportHeight: 700,
-  componentFolder: 'src',
   component: {
+    devServer,
+    componentFolder: 'src',
+    testFiles: '**/*cy-spec.js',
     setupNodeEvents (on, config) {
       on('task', happoTask)
-      devServer(on, config)
 
       // IMPORTANT to return the config object
       // with the any changed environment variables
