+++ npm/react/examples/webpack-options/cypress.config.js
@@ -2,8 +2,7 @@ const { defineConfig } = require('cypress')
 
 // @ts-check
 const path = require('path')
-const { startDevServer } = require('@cypress/webpack-dev-server')
-const babelConfig = require('../../babel.config')
+const babelConfig = require('./babel.config')
 
 module.exports = defineConfig({
   video: false,
